requirements.txt
nose
flake8
coverage
Django
MySQL-python
django-ajax-selects
django-bootstrap3-datetimepicker
django-crispy-forms
django-redis
django-reversion
django-tastypie
django-tastypie-swagger
django-debug-toolbar
django-recaptcha
django-paging
requests
django-axes
python-social-auth
ConfigParser
django-bootstrap3
raven